requirements.txt
asgiref==3.5.2
beautifulsoup4==4.11.1
django-bootstrap4==22.1
django-crispy-forms==1.14.0
Pillow==9.1.1
soupsieve==2.3.2.post1
sqlparse==0.4.2
tzdata==2022.1
djangorestframework-simplejwt==5.2.0
